// src/app/generated/model/attendanceBatchLineRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AttendanceBatchLineRequest { 
    /**
     * Identifier of the worker who was present
     */
    workerId: number;
    /**
     * Clock-in time
     */
    timeIn: string;
    /**
     * Clock-out time; must be after timeIn (400 otherwise)
     */
    timeOut: string;
    /**
     * Optional notes
     */
    notes?: string;
}