// src/app/generated/model/attendanceCalendarEntry.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AttendanceCalendarEntry { 
    /**
     * The date this entry summarizes
     */
    date?: string;
    /**
     * Identifier of the project with recorded attendance on this date
     */
    projectId?: number;
    /**
     * Name of the project with recorded attendance on this date
     */
    projectName?: string;
    /**
     * Number of distinct workers with an attendance record for this project on this date
     */
    workerCount?: number;
}